+++ app/models/moneytree/transaction.rb
@@ -1,0 +1,4 @@
+module Moneytree
+  class Transaction < ApplicationRecord
+  end
+end