+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Master
 
+- Add moneytree transactions
+-
+
 ## 0.1.3 (2020-10-08)
 
 - Fix account to be `null: true` since it is a has_one relation